Stabilize property pages by extracting server actions
